Move charmrevisionmanifold sub-package up to parent for consistency
This is simpler, and consistent with the placement of other worker
Manifold types. Not sure why it was moved into a sub-package in
https://github.com/juju/juju/pull/3642/commits/09799fa54e1d3f2a9e3697d52af1843b4854dec5
-- it doesn't have an import cycle or anything suspicious.
